Don't use mipmaps/gl3w with OpenGL 2
