+++ src-ui/imgui/imgui_image.cpp
@@ -22,17 +22,6 @@ void funcUpdateImageTexture(unsigned int gl_text, uint32_t *buffer, int width, i
     glBindTexture(GL_TEXTURE_2D, 0);
 }
 
-void funcUpdateMMImageTexture(unsigned int gl_text, uint32_t* buffer, int width, int height)
-{
-    glBindTexture(GL_TEXTURE_2D, gl_text);
-    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR_MIPMAP_LINEAR);
-    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
-    glPixelStorei(GL_UNPACK_ROW_LENGTH, 0);
-    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA, width, height, 0, GL_RGBA, GL_UNSIGNED_BYTE, buffer);
-    glGenerateMipmap(GL_TEXTURE_2D);
-    glBindTexture(GL_TEXTURE_2D, 0);
-}
-
 void funcDeleteImageTexture(unsigned int /*gl_text*/)
 {
     //glDeleteTextures(1, &gl_text);
@@ -42,6 +31,6 @@ void bindImageTextureFunctions()
 {
     makeImageTexture = funcMakeImageTexture;
     updateImageTexture = funcUpdateImageTexture;
-    updateMMImageTexture = funcUpdateMMImageTexture;
+    updateMMImageTexture = funcUpdateImageTexture;
     deleteImageTexture = funcDeleteImageTexture;
 }
