+++ src-ui/main.cpp
@@ -1,5 +1,8 @@
 #include <signal.h>
 #include <filesystem>
+#include "imgui/imgui.h"
+#include "imgui/imgui_impl_glfw.h"
+#include "imgui/imgui_impl_opengl3.h"
 #include "gl.h"
 #include "logger.h"
 #include "core/style.h"
@@ -44,10 +47,10 @@ void window_content_scale_callback(GLFWwindow *, float xscale, float)
 }
 
 void bindImageTextureFunctions();
+void bindMMImageTextureFunctions();
 
 int main(int argc, char *argv[])
 {
-    bindImageTextureFunctions();
     initLogger();
 
     if (argc < 5) // Check overall command
@@ -126,11 +129,10 @@ int main(int argc, char *argv[])
     glfwSwapInterval(0); // Disable vsync on loading screen - not needed since frames are only pushed on log updates, and not in a loop
                          // Vsync slows down init process when items are logged quickly
 
-    //Init gl3w
-    if (gl3wInit()) {
-        logger->critical("Could not init gl3w! Exiting");
-        exit(1);
-    }
+    //Set up texture functions
+    bindImageTextureFunctions();
+    if (!fallback_gl)
+        bindMMImageTextureFunctions();
 
     // Setup Dear ImGui context
     IMGUI_CHECKVERSION();
